POC use of updates collection
original placement of updates in microsite front matter was error prone so updates are being overhauled/moved to their own collection.
here is a first step to implementing an updates section. two next steps are to add updates into the _prose.yml, including a multiselect for country tag.
So too does the front matter for the _country/*.md files need to be changed - the id tag needs to be changed to code, or country.