update init build and task parser
buildInitialPages now exports a countryYFM.json
this object includes the original YFM for pages.
it also mirrors the structure of the countryTasks.json.
the idea here is next there will be a 'countryYMFmapper.js file that will see what does and does not exist in the countryTasks.json file and add new tasks accordingly
this also can be extended for other imports (say contact list, calendar list)